implement new hooks when needed, and use the createModal from state to make modals so that new hook can destroy them if there's an authorized error
