+++ frontend/src/components/forums/ForumListPage/ForumListPage.tsx
@@ -1,9 +1,8 @@
-import { useState, useEffect, useMemo, useCallback } from 'react';
+import { useEffect, useMemo, useCallback } from 'react';
 import { RouteComponentProps } from 'react-router';
 import {
   IonContent,
   IonHeader,
-  IonModal,
   IonTitle,
   IonToolbar,
   useIonRouter,
@@ -24,6 +23,7 @@ import { useTr } from '../../../hooks/useTr';
 import { AddListHeader } from '../../common/ListHeader';
 import { NewForumForm } from '../forms/NewForumForm';
 import { useForumUpdateMutation } from '../../../hooks/useForumUpsertMutation';
+import { useAppContext } from '../../../hooks/useAppContext';
 
 interface ForumListPageProps
   extends RouteComponentProps<{
@@ -35,10 +35,14 @@ export function ForumListPage({ match }: ForumListPageProps) {
   const router = useIonRouter();
   const { tr } = useTr();
 
+  const {
+    actions: { createModal },
+  } = useAppContext();
+
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [present] = useIonToast();
 
-  const [isOpenModal, setIsOpenModal] = useState<boolean>(false);
+  const { openModal, closeModal } = createModal();
 
   const [getForums, { data: forumsData, error }] = useGetForumsLazyQuery();
 
@@ -87,6 +91,22 @@ export function ForumListPage({ match }: ForumListPageProps) {
     ));
   }, [error, forumsData, handleEdit, handleGoToForumDetail]);
 
+  const handleOpenModal = () => {
+    openModal(
+      <>
+        <IonHeader>
+          <IonToolbar>
+            <IonTitle>{tr('Add New Forum')}</IonTitle>
+          </IonToolbar>
+        </IonHeader>
+        <IonContent className="ion-padding">
+          <NewForumForm onCreated={closeModal} onCancel={closeModal} />
+        </IonContent>
+      </>,
+      'full',
+    );
+  };
+
   return (
     <PageLayout>
       <Caption>{tr('Community')}</Caption>
@@ -94,29 +114,11 @@ export function ForumListPage({ match }: ForumListPageProps) {
       <AddListHeader
         title={tr('Forums')}
         onClick={() => {
-          setIsOpenModal(true);
+          handleOpenModal();
         }}
       />
 
       <CardListContainer>{cardListComs}</CardListContainer>
-
-      <IonModal isOpen={isOpenModal} onDidDismiss={() => setIsOpenModal(false)}>
-        <IonHeader>
-          <IonToolbar>
-            <IonTitle>{tr('Add New Forum')}</IonTitle>
-          </IonToolbar>
-        </IonHeader>
-        <IonContent className="ion-padding">
-          <NewForumForm
-            onCreated={() => {
-              setIsOpenModal(false);
-            }}
-            onCancel={() => {
-              setIsOpenModal(false);
-            }}
-          />
-        </IonContent>
-      </IonModal>
     </PageLayout>
   );
 }
